--- gateway/src/test/java/ru/practicum/shareit/ShareItGatewayTests.java
@@ -1,13 +1,0 @@
-package ru.practicum.shareit;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ShareItGatewayTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
