+++ server/src/test/java/ru/practicum/shareit/ShareItServerTest.java
@@ -1,13 +1,8 @@
 package ru.practicum.shareit;
 
-import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class ShareItTests {
-
-    @Test
-    void contextLoads() {
-    }
+class ShareItServerTest {
 
 }
