+++ server/src/test/java/ru/practicum/shareit/dto/BookingInDtoTest.java
@@ -1,11 +1,11 @@
-package shareit.booking.dto;
+package ru.practicum.shareit.dto;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.json.JsonTest;
 import org.springframework.boot.test.json.JacksonTester;
 import ru.practicum.shareit.Constants;
-import ru.practicum.shareit.booking.dto.BookingItemDto;
+import ru.practicum.shareit.booking.dto.BookingInDto;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
@@ -13,31 +13,29 @@ import java.time.format.DateTimeFormatter;
 import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
 
 @JsonTest
-class BookingItemDtoTest {
+class BookingInDtoTest {
     private final DateTimeFormatter dtf = Constants.DATE_TIME_FORMATTER;
     @Autowired
-    private JacksonTester<BookingItemDto> json;
+    private JacksonTester<BookingInDto> json;
 
     @Test
     void testSerialize() throws Exception {
-        var dto = new BookingItemDto(
+        var dto = new BookingInDto(
                 1L,
-                2L,
                 LocalDateTime.now(),
-                LocalDateTime.now().plusDays(2L)
+                LocalDateTime.now().plusDays(2L),
+                2L
         );
 
         var result = json.write(dto);
-
         assertThat(result).hasJsonPath("$.id");
-        assertThat(result).hasJsonPath("$.bookerId");
         assertThat(result).hasJsonPath("$.start");
         assertThat(result).hasJsonPath("$.end");
-        assertThat(result).hasEmptyJsonPathValue("$.status");
+        assertThat(result).hasJsonPath("$.itemId");
 
         assertThat(result).extractingJsonPathNumberValue("$.id").isEqualTo(dto.getId().intValue());
-        assertThat(result).extractingJsonPathNumberValue("$.bookerId").isEqualTo(dto.getBookerId().intValue());
         assertThat(result).extractingJsonPathStringValue("$.start").startsWith(dto.getStart().format(dtf));
         assertThat(result).extractingJsonPathStringValue("$.end").startsWith(dto.getEnd().format(dtf));
+        assertThat(result).extractingJsonPathNumberValue("$.itemId").isEqualTo(dto.getItemId().intValue());
     }
 }
