+++ server/src/test/java/ru/practicum/shareit/dto/BookingOutDtoTest.java
@@ -1,4 +1,4 @@
-package shareit.booking.dto;
+package ru.practicum.shareit.dto;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
