+++ server/src/test/java/ru/practicum/shareit/dto/CommentDtoTest.java
@@ -1,45 +1,42 @@
-package shareit.request.dto;
+package ru.practicum.shareit.dto;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.json.JsonTest;
 import org.springframework.boot.test.json.JacksonTester;
 import ru.practicum.shareit.Constants;
-import ru.practicum.shareit.request.dto.RequestDto;
+import ru.practicum.shareit.comment.dto.CommentDto;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
-import java.util.ArrayList;
 
 import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
 
 @JsonTest
-class RequestDtoTest {
+class CommentDtoTest {
     private final DateTimeFormatter dtf = Constants.DATE_TIME_FORMATTER;
     @Autowired
-    private JacksonTester<RequestDto> json;
+    private JacksonTester<CommentDto> json;
 
     @Test
     void testSerialize() throws Exception {
-        var dto = new RequestDto(
+        var dto = new CommentDto(
                 1L,
-                "Request description",
-                2L,
-                LocalDateTime.now(),
-                new ArrayList<>()
+                "Comment 1",
+                "User Name 1",
+                LocalDateTime.now()
         );
 
         var result = json.write(dto);
+
         assertThat(result).hasJsonPath("$.id");
-        assertThat(result).hasJsonPath("$.description");
-        assertThat(result).hasJsonPath("$.requestorId");
+        assertThat(result).hasJsonPath("$.text");
+        assertThat(result).hasJsonPath("$.authorName");
         assertThat(result).hasJsonPath("$.created");
-        assertThat(result).hasJsonPath("$.items");
 
         assertThat(result).extractingJsonPathNumberValue("$.id").isEqualTo(dto.getId().intValue());
-        assertThat(result).extractingJsonPathStringValue("$.description").isEqualTo(dto.getDescription());
-        assertThat(result).extractingJsonPathNumberValue("$.requestorId").isEqualTo(dto.getRequestorId().intValue());
+        assertThat(result).extractingJsonPathStringValue("$.text").isEqualTo(dto.getText());
+        assertThat(result).extractingJsonPathStringValue("$.authorName").isEqualTo(dto.getAuthorName());
         assertThat(result).extractingJsonPathStringValue("$.created").startsWith(dto.getCreated().format(dtf));
-        assertThat(result).extractingJsonPathArrayValue("$.items").hasSize(0);
     }
 }
