+++ server/src/test/java/ru/practicum/shareit/dto/ItemWithBookingDtoTest.java
@@ -1,4 +1,4 @@
-package shareit.item.dto;
+package ru.practicum.shareit.dto;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
