+++ server/src/test/java/ru/practicum/shareit/dto/UserDtoTest.java
@@ -1,4 +1,4 @@
-package shareit.user.dto;
+package ru.practicum.shareit.dto;
 
 import org.assertj.core.api.AssertionsForClassTypes;
 import org.junit.jupiter.api.Assertions;
